Simplify header: full width, no bg, menu only via dot
- Full width header with minimal margins
- Remove background, border, shadow
- Remove nav links, keep only logo + orange dot menu
- Align menu dot and close button positions
